move util to a more apropriate package: backend.database
